Move diff functionality into diff*.go
